refactor: migrate snowflake and IP address packages to shared module
Moved `undecided.erp.common.snowflake` and `undecided.erp.common.ipaddress` to `undecided.shared.common`. Updated all related package references and addressed minor formatting inconsistencies to improve modularity and enable reusability. Removed obsolete `StringUtilsTest` class as it was no longer required.
