--- backend/src/main/java/undecided/erp/common/ipaddress/package-info.java
@@ -1,4 +1,0 @@
-/**
- * IpAddressProviderを格納.
- */
-package undecided.erp.common.ipaddress;
