--- backend/src/main/java/undecided/erp/common/snowflake/package-info.java
@@ -1,4 +1,0 @@
-/**
- * SnowflakeId取得ロジックを格納
- */
-package undecided.erp.common.snowflake;
