+++ backend/src/main/java/undecided/shared/common/ipaddress/package-info.java
@@ -1,0 +1,2 @@
+/** IpAddressProviderを格納. */
+package undecided.shared.common.ipaddress;
