+++ backend/src/main/java/undecided/shared/common/snowflake/RollingSnowflakeIdProvider.java
@@ -1,4 +1,4 @@
-package undecided.erp.common.snowflake;
+package undecided.shared.common.snowflake;
 
 import java.util.ArrayList;
 import java.util.Arrays;
