+++ backend/src/main/java/undecided/shared/common/snowflake/package-info.java
@@ -1,0 +1,2 @@
+/** SnowflakeId取得ロジックを格納 */
+package undecided.shared.common.snowflake;
