+++ backend/src/test/java/undecided/shared/common/snowflake/FixedSnowflakeIdValueProviderTest.java
@@ -1,4 +1,4 @@
-package undecided.erp.common.snowflake;
+package undecided.shared.common.snowflake;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
